hpa_meshing_package/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hpa-meshing"
version = "0.1.0"
description = "Meshing automation package for human-powered aircraft components using Gmsh and SU2 export."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "Project Owner + Codex"}
]
dependencies = [
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "rich>=13.7",
]
optional-dependencies.dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
]

[project.scripts]
hpa-mesh = "hpa_meshing.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
